requirements.txt
PyYAML>=6.0,<7.0
jsonschema>=4.0,<5.0
